Update all allocator constructors to take additional parameters (other than num_cores). These are: optimization target, speedup model, core power, and uncore power. These changes (and changes to any constructed objects) are the majority of this large commit.
Add a generic OptimizeForTarget() function in BaseSpeedupModel. This uses the
constructor parameter opt_target to determine whether to optimize for throughput
or energy, so it can be set at runtime.

Add two new knobs in Zesto for optimization target and speedup model. Not yet
added are core and uncore power knobs.

Update unit tests for the new constructor signatures and functions.

Update makefile.

This is preparation for integrating the new speedup models into the existing
allocator interfaces.