Fixed unit test failure due to different permutation of valid solution.
On some machines with different libstdcxx versions, when multiple processes
have the same scaling factors, any permutation of their core allocations is
still a valid solution. What was happening was that in a test case where all
three processes were identical, the hardcoded solution was (11, 11, 10) but the
solution returned was (10, 11, 11).

Change-Id: I4076ce5b4a7fb53016f9c4c46ce10fece45f3e8f