Clean up BufferManager implementation.
Mostly, got rid of limiting the size, and all the crazienss associated
with reserving hadnshakes and the pool.
locks_[tid] were getting heavily contended in the previous implementation,
and that was limitng performance.
Now, we only grab locks at the smallest granularity (when publishing changes
to fileBuffer_), and wait on condition for fileBuffer_ entries to come.
So, no contention any more.

Also, a few minor cleanups:
- XIOSIM_LOCK is now BasicLockable and movable.
- no need for ReleaseHandshake anymore (which means we can kill callbacks.*).
- we prefer emplacing in the shared maps now.

Change-Id: I113209c0529d444ccbaa40a756e3d4508bd54adf