Expression evaluator for formula statistics.
The evaluator extensively uses unique_ptr to manage all the pointer
passing. Expression objects are the owner of all pointers. Although
ExpressionWrapper is exclusively responsible for calling new, when a new
Expression object is created (either SingleExpression or
CompoundExpression), ownership of the pointer is transferred from
ExpressionWrapper to Expression. Ultimately, this expression is assigned
to a Formula object, so the Formula object is the owner of the root
Expression object.

Tested to work with combinations of two-input and three input operand
functions, so it should scale to above that.

Change-Id: Ic05dc81c24037422a19b0f8bd272162906d5cab8