Send all stats testing printfs to temp files and diff them per test.
This means we don't have to manually check if printouts are correct, and
those printouts no longer pollute the output when running all the unit
tests.

This actually caught a bug in Distribution::print_value() because we
used a float specifier to print an unsigned integer, which caused
undefined behavior and divergence between the Jenkins VM and our
development machines.

Also, I've switched some of the REQUIRE macros to CHECK macros. We don't
need to immediately fail the test if one of checks fails, because it
can be more expedient to see how many of all the checks fail and then
fix multiple bugs at once instead of one at a time. This only applies
for orthogonal checking assertions, of course.

Change-Id: Ic25b259bbac971720ecf6666b2a0daaa31984008