Added missing stats features like saving and accumulates.
In adding these features, StatisticCommon was significantly revamped. It
is now purely an interface, with Statistic<V> implementing all of the
functionality. This was required due to string-type statistics, for
which we typically cannot store pointers to.

Some functions in the new expression evaluator were extracted into .cpp
implementation files due to linking issues.

.clang-format was updated with some tweaks.

BUG: XIOSIM-2, XIOSIM-27

Change-Id: Iac7f5171c51ef928904ad44c4328e131d76bd7eb