Add support for incremental building of Formulas.
Instead of having to construct the complete expression at compile time,
we can now incrementally construct it through the += operator on
Formula. This is required to build equations for which the number of
terms is unknown at compile time (for instance, total instructions
executed by all cores).

Added the exponentiation operator as well.

Change-Id: Icd591757d0d7df85136827d914506de1ebef8336