Initial purge of existing stats library.
This commit replaces the stats library with the new one we have written
and adds an API on top of it to be almost exactly compatible with the
original library, so that we can minimize code changes. The API should
be functional for all statistics except for distributions and formulas.

This will compile, but it will fail all integration tests.

BUG: XIOSIM-27
Change-Id: I8c5619141b7c488e25e11faf4664e40a46bfe6c1