Re-enable speculation. Modify tests.
Result of a quite long performance regression investigation.
The "golden" values in the regression suite were quite wrong,
due to 9fa0272, which had completely messed up speculation before
the oracle rewrite. The branch MPKI values in the golden runs were
unnaturally wrong because we had essentially ignored speculation.

Also, update N.cfg to use RAS-multistack.

TODO: When we have non-trivial branching (fib), we still get a lot
of discrepancy between the feeder predictor and the one in the core
(oracle_nop_handshakes measures that). Doing nested speculation in
the feeder might be one way to resolve that.

Change-Id: Idc52c88cb7fa44116f3cb7b88409da9d34a64501