gettimeofday() returns simulated time, not host time.
The first time a process calls gettimeofday(), we return the host time.
After this, all calls return the simulated time plus this initial
offset.

Note: weird things will happen if we straddle an ROI entry/exit point
with a pair of these calls. We'll handle this case later.

TESTED=integration

Change-Id: Id81c583a386e074b93bb117c4b40583d57baf63a