feeder: Don't track first instrumentation.
Instead, just use the valid bit in the producer buffer to determine
if we need a new handshake.

The bug that triggered this looked like so:
- a start pinpoint at the beginning of a trace;
- a profiling point near the end of *the same* trace;
- starting the pinpoint adds instrumentation to *the next trace*!
- the profiling instrumentation still triggers (that's actually
correct after the call to StartSimSlice);
- we're left with an empty invalid handshake because we never did
the typical instrumentation for the profiling point.
The culprit is starting from the next trace, which we don't have
control over. So, we'll just make the buffers aware of the valid
bit and use that to track if need a new handshake, not explicitly
ask for one on the first instrumentation call.

Change-Id: I7a591a08ca73a51a490cd8c79d1e53ac8fa4be6c