Sample load/store misses by the PC of the instruction.
Change the sampling frequency through the cache_miss_sample_parameter
parameter in the system configuration. Set it to 0 to disable the stat
altogether. Then set the sample_misses parameter in the configuration of
the specific cache of interest to true. Only the caches with
sample_misses true will have the stat appear in the final stat printout.

TESTED=integration

Change-Id: I21b197fbb0d69f14f8f6faff274183812167e9e5