Realistic producer-side implementation of size class.
We can get away without a consumer-side reimplementation because all we
need to do for correct behavior is to take the right code path. However,
this does mean we cannot model speculative accesses and updates to the
size class cache.

Also, there is no way to get stats out of the size class cache (yet).

Change-Id: I3f0224d9fe3aad6c4b199e8280f3269717bee55a