Implemented realistic free list head cache for push and pop.
The head pointer was added as an additional field in the size class
cache. This cache lookup and update required another pair of magic
instructions (shrd and shrx).

Notably, shrx, the cache update instruction, can take a memory operand.
So we use the memory iform on the Pop path to prefetch the next head,
but on the Push path, we use the register iform to simply push the next
head onto the cache, if there is already an entry available for it.  We
use the same software fallback technique as the size class if the cache
lookup misses.

Push and Pop knobs have been combined into ONE knob - if we only
simulate one of them, we get cache coherency problems (e.g. push doesn't
update the free list head in the cache so the next pop gets the old
head, not the new head).

TESTED=manual

Change-Id: I00d1e421085535b13304649545b636255ff7e89a
(cherry picked from commit abc596bba3b490e21a20511fdde6857963507a81)