Allow the size class cache to run with just head pointers.
Usually, we run with both realistic size class and realistic linked list
head pointers because we rely on the size class instructions to create
the cache entries that the head pointers will populate/update. But since
it's useful to be able to isolate the performance benefits of these
optimizations, this mode runs the size class cache with just head
pointers and their corresponding size classes.

To enable this mode, build with -DSIZE_CLASS_CACHE_SLL_ONLY. Be careful
when enabling this mode AND running with realistic size class, since
calling head_update() can now evict valid size class entries!

This commit also restructures size class cache access methods.  This
consolidates the locations where the cache and LRU chains can be
accessed and modified. We strictly enforce the invariant that the LRU
chain has the same length as the number of elements in the cache. Any
lookup to an element in the cache that hits will update the LRU state.

Change-Id: I2be711d05a69ea61faf7071e6e24f3b5a1bb5bee
(cherry picked from commit 451a2eefeae7ea066016658d22c82db1777b7865)