Bring back the (now) correct TLB-finished check.
(after 106220).
Now that we send loads correctly to the TLB, make sure the # of TLB
ports follows published data (2 for NHM and 3 for HSW). We model 2
and 4 because we use a simple hash to determine bank conflicts.

Also fix the page table pretend address calculation so that all
TLB requests don't end up at the same bank.

Change-Id: I0a374432ba7a84df21e52d1d90a7d67de7322d44